seperate getFunction & postFunction
